Adds support to comma separtor

+++ app/src/main/kotlin/com/fibelatti/pigbank/presentation/goals/adapter/GoalsDelegateAdapter.kt
@@ -50,8 +50,8 @@ class GoalsDelegateAdapter @Inject constructor() :
 
                     itemView.textViewDescription.text = description
                     itemView.textViewSavingsProgress.text = itemView.context.resources.getString(R.string.goal_saved_relative, totalSaved, cost)
-                    itemView.progressBarPercent.progress = percentSaved.toInt()
-                    itemView.textViewSavingsProgressPercent.text = itemView.context.resources.getString(R.string.goal_saved_percent, percentSaved.toInt())
+                    itemView.progressBarPercent.progress = percentSaved
+                    itemView.textViewSavingsProgressPercent.text = itemView.context.resources.getString(R.string.goal_saved_percent, percentSaved)
                 }
             }
